Added knobs to switch isosurface depth and others.

+++ share/shaders/self-shadowing/approximate_deep_shadows.glsl
@@ -21,7 +21,7 @@ float approximate_deep_shadow(float shadow_depth, float light_depth, float stran
 }
 
 // Instead of "guessing" the amount of strands in the way, we can find the amount from the strand voxelization.
-float volume_approximated_deep_shadows(sampler3D volume, vec3 strand_position, vec3 light_position, uint steps,
+float volume_approximated_deep_shadows(sampler3D volume, vec3 strand_position, vec3 light_position, float steps,
                                        float strand_alpha, vec3 volume_origin, vec3 volume_size) { // Vol-ADSM.
     float strands = 0;
     float step_size = 1.0f / steps; // for raymarch.
