+++ share/shaders/volume/volume_rendering.glsl
@@ -20,7 +20,7 @@ vec3 volume_tangent(sampler3D volume, vec3 surface, vec3 volume_origin, vec3 vol
 }
 
 // Finds the surface of a volume with at least 'surface_density' starting from 'volume_start' to 'volume_end' when it has been sampled 'step' times.
-vec4 volume_surface(sampler3D volume, vec3 volume_start, vec3 volume_end, uint steps, float surface_density, vec3 volume_origin, vec3 volume_size) {
+vec4 volume_surface(sampler3D volume, vec3 volume_start, vec3 volume_end, float steps, float surface_density, vec3 volume_origin, vec3 volume_size) {
     float density = 0.0f; // current density values.
     float step_size = 1.0f / steps; // for raymarch.
     for (float t = 0.0f; t < 1.0f; t += step_size) {
